+++ flightlib/src/envs/vec_env.cpp
@@ -76,8 +76,8 @@ bool VecEnv<EnvBase>::reset(Ref<MatrixRowMajor<>> obs) {
   return true;
 }
 template<typename EnvBase>
-bool VecEnv<EnvBase>::zero_reset(Ref<MatrixRowMajor<>> obs, const int x_pos,
-  const int y_pos, const int z_pos) {
+bool VecEnv<EnvBase>::zero_reset(Ref<MatrixRowMajor<>> obs, const float x_pos,
+  const float y_pos, const float z_pos) {
   if (obs.rows() != num_envs_ || obs.cols() != obs_dim_) {
     logger_.error(
       "Input matrix dimensions do not match with that of the environment.");
